pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[project]
name = "epigabm"
version = "0.1.0"
description = "Generative agent-based model of influenza spread with LLM-driven behavior"
# Python 3.13+ is not supported because hmmlearn has no prebuilt wheels for 3.13/3.14 on Windows.
requires-python = ">=3.11,<3.13"
dependencies = [
    "numpy>=1.24",
    "pandas>=2.0",
    "scipy>=1.10",
    "loguru>=0.7",
    "pyyaml>=6.0",
    "pydantic>=2.0",
    "tqdm>=4.65",
    "matplotlib>=3.7",
    "tenacity>=8.2",
    "scikit-learn>=1.3",
]

[project.optional-dependencies]
llm = [
    "openai>=1.0",
    "tiktoken>=0.5",
    "ollama>=0.1",
]
viz = [
    "streamlit>=1.30",
    "folium>=0.15",
    "plotly>=5.18",
]
calibration = [
    "hmmlearn>=0.3",
    "SALib>=1.4",
    "tensorflow>=2.15",
    "joblib>=1.3",
]
geo = [
    "geopandas>=0.14",
    "osmnx>=1.9",
]
dev = [
    "pytest>=7.4",
    "ruff>=0.1",
    "ipykernel>=6.25",
]
all = ["epigabm[llm,viz,calibration,geo,dev]"]

[tool.poetry]
packages = [{ include = "src" }]

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
numpy = ">=1.24"
pandas = ">=2.0"
scipy = ">=1.10"
loguru = ">=0.7"
pyyaml = ">=6.0"
pydantic = ">=2.0"
tqdm = ">=4.65"
matplotlib = ">=3.7"
tenacity = ">=8.2"
scikit-learn = ">=1.3"
openai = { version = ">=1.0", optional = true }
tiktoken = { version = ">=0.5", optional = true }
ollama = { version = ">=0.1", optional = true }
streamlit = { version = ">=1.30", optional = true }
folium = { version = ">=0.15", optional = true }
plotly = { version = ">=5.18", optional = true }
hmmlearn = { version = ">=0.3", optional = true }
SALib = { version = ">=1.4", optional = true }
tensorflow = { version = ">=2.15", optional = true }
joblib = { version = ">=1.3", optional = true }
geopandas = { version = ">=0.14", optional = true }
osmnx = { version = ">=1.9", optional = true }

[tool.poetry.extras]
llm = ["openai", "tiktoken", "ollama"]
viz = ["streamlit", "folium", "plotly"]
calibration = ["hmmlearn", "SALib", "tensorflow", "joblib"]
geo = ["geopandas", "osmnx"]
all = ["openai", "tiktoken", "ollama", "streamlit", "folium", "plotly", "hmmlearn", "SALib", "tensorflow", "joblib", "geopandas", "osmnx"]

[tool.poetry.group.dev.dependencies]
pytest = ">=7.4"
ruff = ">=0.1"
ipykernel = ">=6.25"

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
# S, E, I, R are standard SEIR compartment names.
ignore = ["E741"]

[tool.pytest.ini_options]
testpaths = ["tests"]
